+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes
 
 ### Unreleased
 
+* 2017-11-23 - Removed support for Moodle pix icons.
 * 2017-11-23 - Small fix to prevent icons with target blank from being shown when they do not match the language setting.
 * 2017-11-23 - Check compatibility for Moodle 3.3, no functionality change.
 * 2017-11-08 - Updated travis.yml to use newer node version for fixing TravisCI error.